LayoutTests/fast/text/FontFaceSet-check-after-style-update-expected.txt

This function makes sure document.fonts.check() is affected by style changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fonts.check('12px ThisIsNotARealFont') is true
PASS fonts.check('12px WebFont') is false
PASS fonts.check('12px WebFont') is false
PASS fonts.check('12px WebFont') is true
PASS successfullyParsed is true

TEST COMPLETE
